Update Profiles and Reports CRUD's reponses

+++ back/DTOs/ReportDto.cs
@@ -5,7 +5,7 @@ namespace back.Entities
         public string Id { get; set; }
         public DateTime Date { get; set; }
         public int Performance { get; set; }
-        public int AchivedGoals { get; set; }
+        public double AchivedGoals { get; set; }
         public double SavedMoney { get; set; }
         public string ProfileId { get; set; }
         public string Comments { get; set; }
